requirements.txt
# Web server
flask>=3.0.0
gunicorn>=21.2.0

# Core geospatial
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
shapely>=2.0.0
pyproj>=3.6.0
rtree>=1.1.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Visualization
folium>=0.15.0
branca>=0.7.0
matplotlib>=3.7.0

# Machine learning (scoring)
scikit-learn>=1.3.0

# Agent / scheduling
schedule>=1.2.0

# Utilities
python-dotenv>=1.0.0
tabulate>=0.9.0
